requirements.txt
Django==2.0.13
djangorestframework==3.10.3
pytz==2019.2
xlrd==1.2.0
